refactor(portcullis-core): extract Operation/SinkClass/is_exfil to ifc_ops kernel module (MVK M1b)
Move the kernel's operation/sink vocabulary out of the unfenced lib.rs root into a
new ifc_ops.rs kernel module (491 LOC): Operation + LatticeOperation (with its
sealed `private` trait), SinkClass + required_authority/required_integrity +
default_sink_class, and is_exfil_operation. Re-exported at the crate root
(pub use ifc_ops::*), so portcullis_core::{Operation, SinkClass, ...} resolve as
before — zero consumer churn.

- is_exfil_operation reimplemented as a DIRECT match on Operation (no longer routed
  through lib.rs's classify_operation/ExposureLabel exposure machinery), so the
  kernel does not depend on the exposure code. Pinned to the old classifier-based
  definition by the exhaustive is_exfil_operation_matches_classifier test.
- Closes audit finding H1's substance: ROOT_RESIDUALS is now EMPTY — the kernel
  references no crate-root item defined outside a kernel file. ifc_ops added to
  KERNEL_FILES + KERNEL_MODULES; Operation/SinkClass/is_exfil_operation moved from
  ROOT_RESIDUALS to KERNEL_ROOT_PRIMITIVES. Verified non-vacuous (catches an
  injected crate::witness in ifc_ops).

No behavior change: 755 lib tests pass, all-features build clean, clippy clean,
ratchet green.
